@@ -1,66 +1,67 @@
 {
-  "name": "AngularStudy",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@jaspero/ng-confirmations": "^0.4.7",
-    "ag-grid": "^17.0.0",
-    "ag-grid-angular": "^17.0.0",
-    "ag-grid-enterprise": "^17.0.0",
-    "angular-baidu-maps": "^1.0.1",
-    "angular-tree-component": "^7.1.0",
-    "animate.css": "^3.5.2",
-    "bootstrap": "^4.0.0",
-    "core-js": "^2.5.1",
-    "font-awesome": "^4.7.0",
-    "jquery": "^3.3.1",
-    "ng-sidebar": "^6.0.5",
-    "ng-snotify": "^4.2.0",
-    "ng-zorro-antd": "^1.1.0",
-    "ng2-dnd": "^5.0.2",
-    "ng2-semantic-ui": "^0.9.7",
-    "ngx-bootstrap": "^2.0.3",
-    "rxjs": "^5.5.5",
-    "screenfull": "^3.3.2",
-    "zone.js": "^0.8.18"
-  },
-  "devDependencies": {
-    "@angular/cli": "^1.6.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@types/jasmine": "~2.6.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~4.0.1",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.0",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.6.0"
-  }
+   "name": "AngularStudy",
+   "version": "0.0.0",
+   "license": "MIT",
+   "scripts": {
+      "ng": "ng",
+      "start": "ng serve",
+      "build": "ng build",
+      "test": "ng test",
+      "lint": "ng lint",
+      "e2e": "ng e2e"
+   },
+   "private": true,
+   "dependencies": {
+      "@angular/animations": "^6.0.0",
+      "@angular/common": "^6.0.0",
+      "@angular/compiler": "^6.0.0",
+      "@angular/core": "^6.0.0",
+      "@angular/forms": "^6.0.0",
+      "@angular/http": "^6.0.0",
+      "@angular/platform-browser": "^6.0.0",
+      "@angular/platform-browser-dynamic": "^6.0.0",
+      "@angular/router": "^6.0.0",
+      "@jaspero/ng-confirmations": "^0.4.7",
+      "ag-grid": "^17.0.0",
+      "ag-grid-angular": "^17.0.0",
+      "ag-grid-enterprise": "^17.0.0",
+      "angular-baidu-maps": "^1.0.1",
+      "angular-tree-component": "^7.1.0",
+      "animate.css": "^3.5.2",
+      "bootstrap": "^4.0.0",
+      "core-js": "^2.5.4",
+      "font-awesome": "^4.7.0",
+      "jquery": "^3.3.1",
+      "ng-sidebar": "^6.0.5",
+      "ng-snotify": "^4.3.1",
+      "ng-zorro-antd": "^1.1.0",
+      "ng2-dnd": "^5.0.2",
+      "ng2-semantic-ui": "^0.9.7",
+      "ngx-bootstrap": "^2.0.3",
+      "rxjs": "^6.0.0",
+      "screenfull": "^3.3.2",
+      "zone.js": "^0.8.26",
+      "rxjs-compat": "^6.0.0-rc.0"
+   },
+   "devDependencies": {
+      "@angular/compiler-cli": "^6.0.0",
+      "@angular-devkit/build-angular": "~0.6.0",
+      "typescript": "~2.7.2",
+      "@angular/cli": "~6.0.0",
+      "@angular/language-service": "^6.0.0",
+      "@types/jasmine": "~2.8.6",
+      "@types/jasminewd2": "~2.0.3",
+      "@types/node": "~8.9.4",
+      "codelyzer": "~4.2.1",
+      "jasmine-core": "~2.99.1",
+      "jasmine-spec-reporter": "~4.2.1",
+      "karma": "~1.7.1",
+      "karma-chrome-launcher": "~2.2.0",
+      "karma-coverage-istanbul-reporter": "~1.4.2",
+      "karma-jasmine": "~1.1.1",
+      "karma-jasmine-html-reporter": "^0.2.2",
+      "protractor": "~5.3.0",
+      "ts-node": "~5.0.1",
+      "tslint": "~5.9.1"
+   }
 }
