@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {HttpEvent, HttpInterceptor, HttpHandler, HttpRequest} from '@angular/common/http';
-import {Observable} from "rxjs/Observable";
+import {Observable} from "rxjs";
 
 
 @Injectable()
