@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
-import {BehaviorSubject} from 'rxjs/BehaviorSubject';
+import {BehaviorSubject, Observable} from 'rxjs';
 import {MenuData} from '../../model/mainModel';
-import {Observable} from 'rxjs/Observable';
 import {share} from 'rxjs/operators';
 
 @Injectable()
