@@ -1,7 +1,6 @@
 import {Injectable} from '@angular/core';
-import {Subject} from "rxjs/Subject";
+import {Subject, Observable} from "rxjs";
 import {ToastConfig, ToastType} from "../model/toast-model";
-import {Observable} from "rxjs/Observable";
 import {ConfirmationService, ConfirmSettings} from "@jaspero/ng-confirmations";
 
 /**
